requirements.txt
streamlit
numpy
joblib
google-genai
python-dotenv
plotly